Refs #124. Moved hyperlink testing to its own module
